Warn on decompression errors Refactor ThingFilter code Use MpVersion.IsDebug instead of preprocessor directives
