+++ Source/Client/Multiplayer.csproj
@@ -44,7 +44,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\MultiplayerLoader\MultiplayerLoader.csproj" ExcludeAssets="runtime" Private="false" />
-    <ProjectReference Include="..\Common\Common.csproj" ExcludeAssets="runtime"  />
+    <ProjectReference Include="..\Common\Common.csproj" ExcludeAssets="runtime" />
   </ItemGroup>
 
   <ItemGroup>
