+++ Source/MultiplayerLoader/Prepatches.cs
@@ -1,6 +1,7 @@
 ﻿using Mono.Cecil;
 using Mono.Cecil.Cil;
 using MonoMod.Utils;
+using Multiplayer.Common;
 using Prepatcher;
 using RimWorld;
 
@@ -11,9 +12,10 @@ public static class Prepatches
     [FreePatch]
     static void DontSetResolution(ModuleDefinition module)
     {
-#if DEBUG
-        var utilType = module.ImportReference(typeof(ResolutionUtility)).Resolve();
-        utilType.FindMethod("SetResolutionRaw").Body.Instructions.Insert(0, Instruction.Create(OpCodes.Ret));
-#endif
+        if (MpVersion.IsDebug)
+        {
+            var utilType = module.ImportReference(typeof(ResolutionUtility)).Resolve();
+            utilType.FindMethod("SetResolutionRaw").Body.Instructions.Insert(0, Instruction.Create(OpCodes.Ret));
+        }
     }
 }
